refactor: apply suggested changes
- clean up code based on suggestions
- store jobs per dimension
